refactor: split terminal display extensions into terminal/result.rs
`OutcomeDisplay` and `RunResultRender` are both display-extension
traits on dag domain types — they belong together in a sibling file
rather than bloating `terminal.rs`. Moved both into
`cli/src/terminal/result.rs` (singular, matching `metrics/result.rs`
on the dag side). `terminal.rs` stays focused on the `Terminal`
facade + ANSI constants + `stderr_supports_color`.

Along the way, two small fixes in the table module:

- `ReplicaRow::for_result` was calling
  `metrics.committed_leaders(authority)`, which filters to leaders
  that replica itself proposed; the table column wants the total
  count across all leader authorities. Switched to
  `RunResult::commit_count_per_replica` (the same source as the
  suite-row numbers).
- The redundant `duration_secs` parameter on `ReplicaRow::for_result`
  is gone — `result.duration.as_secs()` has the same value.

Tried migrating `for_config` / `for_result` to `From<&X>` impls but
the orphan rule forbids `impl From<&SimulationConfig> for Vec<ConfigRow>`
— `Vec` isn't fundamental, so a local type parameter doesn't count
for coherence. Named associated functions stay.

Tracked by #66.
